test: DCI 패턴을 적용한 CouponTest 적용
