+++ backend/src/test/java/com/woowacourse/kkogkkog/domain/CouponTest.java
@@ -7,199 +7,276 @@ import com.woowacourse.kkogkkog.exception.ForbiddenException;
 import com.woowacourse.kkogkkog.exception.InvalidRequestException;
 import com.woowacourse.kkogkkog.exception.coupon.SameSenderReceiverException;
 import com.woowacourse.kkogkkog.fixture.MemberFixture;
-import java.util.stream.Stream;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 
-@SuppressWarnings("NonAsciiCharacters")
+@DisplayName("Coupon 클래스의")
 public class CouponTest {
 
-    @Test
-    void 쿠폰을_생성할_수_있다() {
-        Member sender = MemberFixture.ROOKIE;
-        Member receiver = MemberFixture.ARTHUR;
+    @Nested
+    @DisplayName("생성자는")
+    class Constructor {
 
-        Coupon actual = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
-            CouponType.COFFEE,
-            CouponStatus.READY);
+        @Test
+        @DisplayName("쿠폰의 정보를 받으면, 쿠폰을 생성한다.")
+        void success() {
+            Member sender = MemberFixture.ROOKIE;
+            Member receiver = MemberFixture.ARTHUR;
 
-        assertThat(actual).isNotNull();
+            Coupon actual = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                CouponType.COFFEE, CouponStatus.READY);
+
+            assertThat(actual).isNotNull();
+        }
+
+        @Test
+        @DisplayName("보낸 사람과 받는 사람이 동일하면, 예외를 던진다.")
+        void fail_sameSenderAndReceiver() {
+            Member member = MemberFixture.ROOKIE;
+            assertThatThrownBy(
+                () -> new Coupon(null, member, member, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.READY)
+            ).isInstanceOf(SameSenderReceiverException.class);
+        }
     }
 
-    @Test
-    void 보낸_사람과_받는_사람이_같은_경우_예외를_던진다() {
-        Member member = MemberFixture.ROOKIE;
-        assertThatThrownBy(
-            () -> new Coupon(null, member, member, "한턱쏘는", "추가 메세지", "#241223", CouponType.COFFEE,
-                CouponStatus.READY)
-        ).isInstanceOf(SameSenderReceiverException.class);
-    }
+    @Nested
+    @DisplayName("changeStatus 메서드는")
+    class ChangeStatus {
 
-    @Test
-    void 받은_사람은_READY_상태의_쿠폰에_대한_사용_요청을_보낼_수_있다() {
-        Member sender = MemberFixture.ROOKIE;
-        Member receiver = MemberFixture.ARTHUR;
-        Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
-            CouponType.COFFEE,
-            CouponStatus.READY);
+        @Nested
+        @DisplayName("READY 상태의 쿠폰에")
+        class Ready {
 
-        coupon.changeStatus(CouponEvent.REQUEST, receiver);
+            @Test
+            @DisplayName("받은 사람이 REQUEST 를 보내면, REQUESTED 로 변경한다.")
+            void success_request() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.READY);
 
-        assertThat(coupon.getCouponStatus()).isEqualTo(CouponStatus.REQUESTED);
-    }
+                coupon.changeStatus(CouponEvent.REQUEST, receiver);
 
-    @Test
-    void 보낸_사람은_쿠폰_사용_요청을_보낼_수_없다() {
-        Member sender = MemberFixture.ROOKIE;
-        Member receiver = MemberFixture.ARTHUR;
-        Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
-            CouponType.COFFEE,
-            CouponStatus.READY);
+                assertThat(coupon.getCouponStatus()).isEqualTo(CouponStatus.REQUESTED);
+            }
 
-        assertThatThrownBy(() -> coupon.changeStatus(CouponEvent.REQUEST, sender))
-            .isInstanceOf(ForbiddenException.class);
-    }
+            @Test
+            @DisplayName("보낸 사람이 REQUEST 를 보내면, 예외를 던진다.")
+            void fail_requestBySender() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.READY);
 
-    @Test
-    void 받은_사람은_REQUESTED_상태의_쿠폰에_대한_사용_요청_취소를_보낼_수_있다() {
-        Member sender = MemberFixture.ROOKIE;
-        Member receiver = MemberFixture.ARTHUR;
-        Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
-            CouponType.COFFEE,
-            CouponStatus.REQUESTED);
+                assertThatThrownBy(() -> coupon.changeStatus(CouponEvent.REQUEST, sender))
+                    .isInstanceOf(ForbiddenException.class);
+            }
 
-        coupon.changeStatus(CouponEvent.CANCEL, receiver);
+            @Test
+            @DisplayName("받은 사람이 CANCEL 을 보내면, 예외를 던진다.")
+            void fail_cancelReady() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.READY);
 
-        assertThat(coupon.getCouponStatus()).isEqualTo(CouponStatus.READY);
-    }
+                assertThatThrownBy(() -> coupon.changeStatus(CouponEvent.CANCEL, receiver))
+                    .isInstanceOf(InvalidRequestException.class);
+            }
 
-    @Test
-    void 받은_사람은_READY_상태의_쿠폰에_대한_사용_요청_취소를_보낼_수_없다() {
-        Member sender = MemberFixture.ROOKIE;
-        Member receiver = MemberFixture.ARTHUR;
-        Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
-            CouponType.COFFEE,
-            CouponStatus.READY);
+            @Test
+            @DisplayName("보낸 사람이 ACCEPT 를 보내면, 예외를 던진다.")
+            void fail_accept() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.READY);
 
-        assertThatThrownBy(() -> coupon.changeStatus(CouponEvent.CANCEL, receiver))
-            .isInstanceOf(InvalidRequestException.class);
-    }
+                assertThatThrownBy(() -> coupon.changeStatus(CouponEvent.ACCEPT, sender))
+                    .isInstanceOf(InvalidRequestException.class);
+            }
 
-    @Test
-    void 보낸_사람은_쿠폰_사용_요청_취소를_보낼_수_없다() {
-        Member sender = MemberFixture.ROOKIE;
-        Member receiver = MemberFixture.ARTHUR;
-        Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
-            CouponType.COFFEE,
-            CouponStatus.REQUESTED);
+            @Test
+            @DisplayName("보낸 사람이 FINISH 를 보내면, FINISHED 로 변경한다.")
+            void success_finish_bySender() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.READY);
 
-        assertThatThrownBy(() -> coupon.changeStatus(CouponEvent.CANCEL, sender))
-            .isInstanceOf(ForbiddenException.class);
-    }
+                coupon.changeStatus(CouponEvent.FINISH, sender);
 
-    @Test
-    void 보낸_사람은_REQUESTED_상태의_쿠폰에_대한_사용_요청_수락을_보낼_수_있다() {
-        Member sender = MemberFixture.ROOKIE;
-        Member receiver = MemberFixture.ARTHUR;
-        Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
-            CouponType.COFFEE,
-            CouponStatus.REQUESTED);
+                assertThat(coupon.getCouponStatus()).isEqualTo(CouponStatus.FINISHED);
+            }
 
-        coupon.changeStatus(CouponEvent.ACCEPT, sender);
+            @Test
+            @DisplayName("받은 사람이 FINISH 를 보내면, FINISHED 로 변경한다.")
+            void success_finish_byReceiver() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.READY);
 
-        assertThat(coupon.getCouponStatus()).isEqualTo(CouponStatus.ACCEPTED);
-    }
+                coupon.changeStatus(CouponEvent.FINISH, receiver);
 
-    @Test
-    void 보낸_사람은_READY_상태의_쿠폰에_대한_사용_요청_수락을_보낼_수_없다() {
-        Member sender = MemberFixture.ROOKIE;
-        Member receiver = MemberFixture.ARTHUR;
-        Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
-            CouponType.COFFEE,
-            CouponStatus.READY);
+                assertThat(coupon.getCouponStatus()).isEqualTo(CouponStatus.FINISHED);
+            }
 
-        assertThatThrownBy(() -> coupon.changeStatus(CouponEvent.ACCEPT, sender))
-            .isInstanceOf(InvalidRequestException.class);
-    }
+            @Test
+            @DisplayName("쿠폰과 관련없는 사람이 FINISH 를 보내면, 예외를 던진다.")
+            void fail_othersNotAllowed() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Member requester = MemberFixture.JEONG;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.READY);
 
-    @Test
-    void 받은_사람은_쿠폰_사용_요청_수락을_보낼_수_없다() {
-        Member sender = MemberFixture.ROOKIE;
-        Member receiver = MemberFixture.ARTHUR;
-        Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
-            CouponType.COFFEE,
-            CouponStatus.REQUESTED);
+                assertThatThrownBy(() -> coupon.changeStatus(CouponEvent.ACCEPT, requester))
+                    .isInstanceOf(ForbiddenException.class);
+            }
+        }
 
-        assertThatThrownBy(() -> coupon.changeStatus(CouponEvent.ACCEPT, receiver))
-            .isInstanceOf(ForbiddenException.class);
-    }
+        @Nested
+        @DisplayName("REQUESTED 상태의 쿠폰에")
+        class Requested {
 
-    @ParameterizedTest
-    @MethodSource("provideSenderAndReceiver")
-    void READY_상태의_쿠폰에_대한_사용_완료를_보낼_수_있다(Member sender, Member receiver, Member requester) {
-        Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
-            CouponType.COFFEE,
-            CouponStatus.READY);
+            @Test
+            @DisplayName("받은 사람이 CANCEL 을 보내면, READY 로 변경한다.")
+            void success_cancel() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.REQUESTED);
 
-        coupon.changeStatus(CouponEvent.FINISH, requester);
+                coupon.changeStatus(CouponEvent.CANCEL, receiver);
 
-        assertThat(coupon.getCouponStatus()).isEqualTo(CouponStatus.FINISHED);
-    }
+                assertThat(coupon.getCouponStatus()).isEqualTo(CouponStatus.READY);
+            }
 
-    @ParameterizedTest
-    @MethodSource("provideSenderAndReceiver")
-    void REQUESTED_상태의_쿠폰에_대한_사용_완료를_보낼_수_있다(Member sender, Member receiver, Member requester) {
-        Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
-            CouponType.COFFEE,
-            CouponStatus.REQUESTED);
+            @Test
+            @DisplayName("보낸 사람이 CANCEL 을 보내면, 예외를 던진다.")
+            void fail_cancel_senderNotAllowed() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.REQUESTED);
 
-        coupon.changeStatus(CouponEvent.FINISH, requester);
+                assertThatThrownBy(() -> coupon.changeStatus(CouponEvent.CANCEL, sender))
+                    .isInstanceOf(ForbiddenException.class);
+            }
 
-        assertThat(coupon.getCouponStatus()).isEqualTo(CouponStatus.FINISHED);
-    }
+            @Test
+            @DisplayName("보낸 사람이 ACCEPT 를 보내면, ACCEPTED 로 변경한다.")
+            void success_accept() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.REQUESTED);
 
-    @ParameterizedTest
-    @MethodSource("provideSenderAndReceiver")
-    void ACCEPTED_상태의_쿠폰에_대한_사용_완료를_보낼_수_있다(Member sender, Member receiver, Member requester) {
-        Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
-            CouponType.COFFEE,
-            CouponStatus.ACCEPTED);
+                coupon.changeStatus(CouponEvent.ACCEPT, sender);
 
-        coupon.changeStatus(CouponEvent.FINISH, requester);
+                assertThat(coupon.getCouponStatus()).isEqualTo(CouponStatus.ACCEPTED);
+            }
 
-        assertThat(coupon.getCouponStatus()).isEqualTo(CouponStatus.FINISHED);
-    }
+            @Test
+            @DisplayName("받은 사람이 ACCEPT 를 보내면, 예외를 던진다.")
+            void fail_accept_receiverNotAllowed() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.REQUESTED);
 
-    @ParameterizedTest
-    @MethodSource("provideSenderAndReceiver")
-    void FINISHED_상태의_쿠폰에_대한_사용_완료를_보낼_수_없다(Member sender, Member receiver, Member requester) {
-        Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
-            CouponType.COFFEE,
-            CouponStatus.FINISHED);
+                assertThatThrownBy(() -> coupon.changeStatus(CouponEvent.ACCEPT, receiver))
+                    .isInstanceOf(ForbiddenException.class);
+            }
 
-        assertThatThrownBy(() -> coupon.changeStatus(CouponEvent.FINISH, requester))
-            .isInstanceOf(InvalidRequestException.class);
-    }
+            @Test
+            @DisplayName("보낸 사람이 FINISH 를 보내면, FINISHED 로 변경한다.")
+            void success_finish_bySender() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.REQUESTED);
 
-    @Test
-    void 보낸_사람과_받은_사람이_아니면_FINISHED_상태의_쿠폰에_대한_사용_완료를_보낼_수_없다() {
-        Member sender = MemberFixture.ROOKIE;
-        Member receiver = MemberFixture.ARTHUR;
-        Member requester = MemberFixture.JEONG;
-        Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
-            CouponType.COFFEE,
-            CouponStatus.READY);
+                coupon.changeStatus(CouponEvent.FINISH, sender);
 
-        assertThatThrownBy(() -> coupon.changeStatus(CouponEvent.ACCEPT, requester))
-            .isInstanceOf(ForbiddenException.class);
-    }
+                assertThat(coupon.getCouponStatus()).isEqualTo(CouponStatus.FINISHED);
+            }
 
-    public static Stream<Arguments> provideSenderAndReceiver() {
-        return Stream.of(
-            Arguments.of(MemberFixture.ROOKIE, MemberFixture.ARTHUR, MemberFixture.ROOKIE),
-            Arguments.of(MemberFixture.ROOKIE, MemberFixture.ARTHUR, MemberFixture.ARTHUR)
-        );
+            @Test
+            @DisplayName("보낸 사람이 FINISH 를 보내면, FINISHED 로 변경한다.")
+            void success_finish_byReceiver() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.REQUESTED);
+
+                coupon.changeStatus(CouponEvent.FINISH, receiver);
+
+                assertThat(coupon.getCouponStatus()).isEqualTo(CouponStatus.FINISHED);
+            }
+        }
+
+        @Nested
+        @DisplayName("ACCEPTED 상태의 쿠폰에")
+        class Accepted {
+
+            @Test
+            @DisplayName("보낸 사람이 FINISH 를 보내면, FINISHED 로 변경한다")
+            void success_finish_bySender() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.ACCEPTED);
+
+                coupon.changeStatus(CouponEvent.FINISH, sender);
+
+                assertThat(coupon.getCouponStatus()).isEqualTo(CouponStatus.FINISHED);
+            }
+
+            @Test
+            @DisplayName("받은 사람이 FINISH 를 보내면, FINISHED 로 변경한다")
+            void success_finish_byReceiver() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.ACCEPTED);
+
+                coupon.changeStatus(CouponEvent.FINISH, receiver);
+
+                assertThat(coupon.getCouponStatus()).isEqualTo(CouponStatus.FINISHED);
+            }
+        }
+
+        @Nested
+        @DisplayName("FINISHED 상태의 쿠폰에")
+        class Finished {
+
+            @Test
+            @DisplayName("보낸 사람이 FINISH 를 보내면, 예외를 던진다.")
+            void fail_finish_bySender() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.FINISHED);
+
+                assertThatThrownBy(() -> coupon.changeStatus(CouponEvent.FINISH, sender))
+                    .isInstanceOf(InvalidRequestException.class);
+            }
+
+            @Test
+            @DisplayName("받은 사람이 FINISH 를 보내면, 예외를 던진다.")
+            void fail_finish_byReceiver() {
+                Member sender = MemberFixture.ROOKIE;
+                Member receiver = MemberFixture.ARTHUR;
+                Coupon coupon = new Coupon(null, sender, receiver, "한턱쏘는", "추가 메세지", "#241223",
+                    CouponType.COFFEE, CouponStatus.FINISHED);
+
+                assertThatThrownBy(() -> coupon.changeStatus(CouponEvent.FINISH, receiver))
+                    .isInstanceOf(InvalidRequestException.class);
+            }
+        }
     }
 }
