+++ backend/src/test/java/com/woowacourse/kkogkkog/fixture/MemberFixture.java
@@ -1,8 +1,6 @@
 package com.woowacourse.kkogkkog.fixture;
 
 import com.woowacourse.kkogkkog.domain.Member;
-import java.util.stream.Stream;
-import org.junit.jupiter.params.provider.Arguments;
 
 public class MemberFixture {
 
@@ -13,10 +11,4 @@ public class MemberFixture {
     public static Member NON_EXISTING_MEMBER = new Member(99999L, "no-one@gmail.com",
         "password1234!", "존재하지_않는_사용자");
 
-    public static Stream<Arguments> provideSenderAndReceiver() {
-        return Stream.of(
-            Arguments.of(MemberFixture.ROOKIE, MemberFixture.ARTHUR, MemberFixture.ROOKIE),
-            Arguments.of(MemberFixture.ROOKIE, MemberFixture.ARTHUR, MemberFixture.ARTHUR)
-        );
-    }
 }
